requests: fix getting state of stages
The paralelization made getting the state of allocation/cleanup stages a
bit more tricky. Reason is that listing SAUs (consequently getting state
of stages) can happen before stages are created. However, when a SAU is
created, we know that corresponding stage must be created; thus from the
user's point of view, it is in `IN_QUEUE` state.