instance_app: utilize size attribute of Pool
Instead of counting the number of SAUs to get the current size of
a pool, use the size attribute.

The import thing is to properly update the size attribute. That
means that upon allocation the size is increased and upon cleanup
decreased. In addition, to avoid race conditions concerning
increase/decrease of size, the transaction.atomic block is needed.