instance_app: fix SAU does not exists
Upon allocation, the SAUs are created during the POST request and then
send to the default worker. But Django does not inserts row
immediatelly; thus default worker fails.

Luckly, the transaction.on_commit saves the day. As argument it takes
the function that will be called after the atomic transaction is done.
For more details see
https://docs.djangoproject.com/en/4.2/topics/db/transactions/#performing-actions-after-commit