instance_app: fix Pool serializer
When migrating from older versions, the size of pool will always be 0.
This is because older versions do not store the size for pool, thus the
default value of 0 is applied. This causes an error during the cleanup,
as the cleanup tries to decrease the size below zero which is restricted
by PositiveIntegerField.