views: fix duble swagger_auto_schema application
When migrating to Django 5.0, I needed to use drf-yasg instead of
drf-yasg2 (drf-yasg is actually newer version). For that reason,
everytime there was a duplicate application of swagget auto schema (on
class and on method) the error was thrown. Thus the endpoint was not
executed. I fixed that, but the problem with swagger not properly
configured for all endpoint persist.