feat: support GitHub as definition provider
The GitHub client has a bit more straightforward interface so it is not
necessary to implement all methods of the parent class (or the
implementation returns inputs).

Besides the GitHubProvider I have updated the GitLab and Definition
providers. The update is only moving some parts of code that are the
same for both providers to the parent class.

Lastly, I have added the configuration example on how to configure both
GitLab and GitHub providers at the same time.

For now only the github.com server is supported. For support of custom
github servers, the subsequent implementations are needed.